access to comment

+++ src/Controller/CommentController.php
@@ -11,10 +11,12 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/comment')]
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+
+#[Route('/comment', name:'comment')]
 class CommentController extends AbstractController
 {
-    #[Route('/', name: 'comment_index', methods: ['GET'])]
+    #[Route('/', name: '_index', methods: ['GET'])]
     public function index(CommentRepository $commentRepository): Response
     {
         return $this->render('comment/index.html.twig', [
@@ -22,7 +24,10 @@ class CommentController extends AbstractController
         ]);
     }
 
-    #[Route('/new', name: 'comment_new', methods: ['GET', 'POST'])]
+    /**
+     * @IsGranted("ROLE_CONTRIBUTOR")
+     **/
+    #[Route('/new', name: '_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $comment = new Comment();
@@ -42,7 +47,7 @@ class CommentController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'comment_show', methods: ['GET'])]
+    #[Route('/{id}', name: '_show', methods: ['GET'])]
     public function show(Comment $comment): Response
     {
         return $this->render('comment/show.html.twig', [
@@ -50,9 +55,18 @@ class CommentController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'comment_edit', methods: ['GET', 'POST'])]
+    /**
+     * @IsGranted("ROLE_CONTRIBUTOR")
+     * */
+    #[Route('/{id}/edit', name: '_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Comment $comment, EntityManagerInterface $entityManager): Response
     {
+        // only the owner can edit the comment
+        if (!($this->getUser() == $comment->getAuthor())) {
+        // If not the owner, throws a 403 Access Denied exception
+        throw new AccessDeniedException('Seul l\'autheur peut supprimer ce commentaire');
+        }
+
         $form = $this->createForm(CommentType::class, $comment);
         $form->handleRequest($request);
 
@@ -68,9 +82,19 @@ class CommentController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'comment_delete', methods: ['POST'])]
+    /**
+     * @IsGranted("ROLE_CONTRIBUTOR")
+     * @IsGranted("ROLE_ADMIN")
+     **/
+    #[Route('/{id}', name: '_delete', methods: ['POST'])]
     public function delete(Request $request, Comment $comment, EntityManagerInterface $entityManager): Response
     {
+        // Check wether the logged in user is the owner of the program
+        if (!($this->getUser() == $comment->getAuthor()) && !$this->isGranted('ROLE_ADMIN')) {
+            // If not the owner, throws a 403 Access Denied exception
+            throw new AccessDeniedException('Seul l\'autheur peut supprimer ce commentaire');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->request->get('_token'))) {
             $entityManager->remove($comment);
             $entityManager->flush();
